Bugfix release

This should, resolve #4

@suhrab Do you mind testing on your environment?

+++ zxcvbn.net/Zxcvbn.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Jurassic;
+
 
 namespace zxcvbn.net
 {
@@ -22,7 +22,7 @@ namespace zxcvbn.net
             return Instance.Entropy(password);
         }
 
-        public static int crack_time(String password)
+        public static double crack_time(String password)
         {
             return Instance.crack_time(password);
         }
